test: bound harness memory ceilings to the session schema maximum
The session validator now bounds u64 schema fields to i64::MAX, so the
harnesses that declared an effectively unbounded memory ceiling with
u64::MAX no longer parse. The value means "do not let the declared
session limit be the binding constraint", so it becomes the schema
maximum i64::MAX as u64.

+++ crates/miso-engine-builtins-compiler/tests/allocation_tracker.rs
@@ -59,7 +59,7 @@ fn session(track_count: u32) -> miso_engine_session::CompiledSession {
     template.dynamic.effects.clear();
     template.simd2.effects.clear();
     model.automation.clear();
-    model.limits.memory_bytes = u64::MAX;
+    model.limits.memory_bytes = i64::MAX as u64;
     model.tracks.clear();
     model
         .tracks
